Fru-device : Extended support for 16bit logic bus
The uint8_t bus of getFruInfo() restricted the number of logical I2C
buses that we could implement in the sysfs interface, and it was
unfortunately static_cast'ed to uint_8.which limit on busses to 255.
while Linux kernel can extend the logical I2C bus number to more than
512 without any issues.
This commit add the support of 16 bit address EEPROM.

reference Link :-
https://lore.kernel.org/all/CACWQX82=MuAavxCqOerxi-Sdywh0xatb-f+1YzGyVSg74oNGqA@mail.gmail.com/T/

Tested :
fru-device successfully recognizes all  8 and 16 bit  FRU EEPROM .

dependent commit:-
https://gerrit.openbmc.org/c/openbmc/phosphor-host-ipmid/+/58635

+++ src/fru_device.cpp
@@ -697,9 +697,9 @@ void addFruObjectToDbus(
             {
                 if (isMuxBus(bus) && bus != busIface.first.first &&
                     address == busIface.first.second &&
-                    (getFRUInfo(static_cast<uint8_t>(busIface.first.first),
+                    (getFRUInfo(static_cast<uint16_t>(busIface.first.first),
                                 static_cast<uint8_t>(busIface.first.second)) ==
-                     getFRUInfo(static_cast<uint8_t>(bus),
+                     getFRUInfo(static_cast<uint16_t>(bus),
                                 static_cast<uint8_t>(address))))
                 {
                     // This device is already added to the lower numbered bus,
@@ -1094,7 +1094,7 @@ bool updateFRUProperty(
     std::vector<uint8_t> fruData;
     try
     {
-        fruData = getFRUInfo(static_cast<uint8_t>(bus),
+        fruData = getFRUInfo(static_cast<uint16_t>(bus),
                              static_cast<uint8_t>(address));
     }
     catch (const std::invalid_argument& e)
@@ -1304,14 +1304,14 @@ int main()
                      objServer, systemBus);
     });
 
-    iface->register_method("ReScanBus", [&](uint8_t bus) {
+    iface->register_method("ReScanBus", [&](uint16_t bus) {
         rescanOneBus(busMap, bus, dbusInterfaceMap, true, unknownBusObjectCount,
                      powerIsOn, objServer, systemBus);
     });
 
     iface->register_method("GetRawFru", getFRUInfo);
 
-    iface->register_method("WriteFru", [&](const uint8_t bus,
+    iface->register_method("WriteFru", [&](const uint16_t bus,
                                            const uint8_t address,
                                            const std::vector<uint8_t>& data) {
         if (!writeFRU(bus, address, data))
